Convert bytearray clear method

+++ boa3_test/tests/test_tuple.py
@@ -260,7 +260,7 @@ class TestTuple(BoaTest):
                 + Opcode.ISTYPE
                 + Type.bytearray.stack_item
                 + Opcode.JMPIFNOT
-                + Integer(8).to_byte_array(signed=True, min_length=1)
+                + Integer(5).to_byte_array(signed=True, min_length=1)
                 + Opcode.CAT
                 + Opcode.JMP
                 + Integer(5).to_byte_array(min_length=1)
@@ -337,7 +337,7 @@ class TestTuple(BoaTest):
                 + Opcode.ISTYPE
                 + Type.bytearray.stack_item
                 + Opcode.JMPIFNOT
-                + Integer(8).to_byte_array(signed=True, min_length=1)
+                + Integer(5).to_byte_array(signed=True, min_length=1)
                 + Opcode.CAT
                 + Opcode.JMP
                 + Integer(5).to_byte_array(min_length=1)
@@ -407,7 +407,7 @@ class TestTuple(BoaTest):
                 + Opcode.ISTYPE
                 + Type.bytearray.stack_item
                 + Opcode.JMPIFNOT
-                + Integer(8).to_byte_array(signed=True, min_length=1)
+                + Integer(5).to_byte_array(signed=True, min_length=1)
                 + Opcode.CAT
                 + Opcode.JMP
                 + Integer(5).to_byte_array(min_length=1)
@@ -501,7 +501,7 @@ class TestTuple(BoaTest):
                 + Opcode.ISTYPE
                 + Type.bytearray.stack_item
                 + Opcode.JMPIFNOT
-                + Integer(8).to_byte_array(signed=True, min_length=1)
+                + Integer(5).to_byte_array(signed=True, min_length=1)
                 + Opcode.CAT
                 + Opcode.JMP
                 + Integer(5).to_byte_array(min_length=1)
